fixed errors
