+++ lab_1/README.md
@@ -1,5 +1,5 @@
 # Основы информационной безопасности
 
-## Лабораторная №1 "Простейшие методы шифрований текстовых сообщений"
+## Лабораторная №1 "Простейшие методы шифрований текстовых сообщений. Шифрование методом простой подстановки"
 
 Дьячков Кирилл Андреевич 6212-100503D
