+++ lab_1/texts/task1/encrypt.txt
@@ -1,54 +1,1 @@
-xf'qf mp tsqbmhfqt sp kpuf
-zpv lmpx sgf qvkft bmc tp cp j (cp j)
-b evkk dpnnjsnfms't xgbs j'n sgjmljmh pe
-zpv xpvkcm's hfs sgjt eqpn bmz psgfq hvz
-j ivts xbmmb sfkk zpv gpx j'n effkjmh
-hpssb nblf zpv vmcfqtsbmc
-mfufq hpmmb hjuf zpv vo
-mfufq hpmmb kfs zpv cpxm
-mfufq hpmmb qvm bqpvmc bmc cftfqs zpv
-mfufq hpmmb nblf zpv dqz
-mfufq hpmmb tbz hppcazf
-mfufq hpmmb sfkk b kjf bmc gvqs zpv
-xf'uf lmpxm fbdg psgfq epq tp kpmh
-zpvq gfbqs't affm bdgjmh, avs zpv'qf spp tgz sp tbz js (tbz js)
-jmtjcf, xf apsg lmpx xgbs't affm hpjmh pm (hpjmh pm)
-xf lmpx sgf hbnf bmc xf'qf hpmmb okbz js
-bmc je zpv btl nf gpx j'n effkjmh
-cpm's sfkk nf zpv'qf spp akjmc sp tff
-mfufq hpmmb hjuf zpv vo
-mfufq hpmmb kfs zpv cpxm
-mfufq hpmmb qvm bqpvmc bmc cftfqs zpv
-mfufq hpmmb nblf zpv dqz
-mfufq hpmmb tbz hppcazf
-mfufq hpmmb sfkk b kjf bmc gvqs zpv
-mfufq hpmmb hjuf zpv vo
-mfufq hpmmb kfs zpv cpxm
-mfufq hpmmb qvm bqpvmc bmc cftfqs zpv
-mfufq hpmmb nblf zpv dqz
-mfufq hpmmb tbz hppcazf
-mfufq hpmmb sfkk b kjf bmc gvqs zpv
-xf'uf lmpxm fbdg psgfq epq tp kpmh
-zpvq gfbqs't affm bdgjmh, avs zpv'qf spp tgz sp tbz js (sp tbz js)
-jmtjcf, xf apsg lmpx xgbs't affm hpjmh pm (hpjmh pm)
-xf lmpx sgf hbnf bmc xf'qf hpmmb okbz js
-j ivts xbmmb sfkk zpv gpx j'n effkjmh
-hpssb nblf zpv vmcfqtsbmc
-mfufq hpmmb hjuf zpv vo
-mfufq hpmmb kfs zpv cpxm
-mfufq hpmmb qvm bqpvmc bmc cftfqs zpv
-mfufq hpmmb nblf zpv dqz
-mfufq hpmmb tbz hppcazf
-mfufq hpmmb sfkk b kjf bmc gvqs zpv
-mfufq hpmmb hjuf zpv vo
-mfufq hpmmb kfs zpv cpxm
-mfufq hpmmb qvm bqpvmc bmc cftfqs zpv
-mfufq hpmmb nblf zpv dqz
-mfufq hpmmb tbz hppcazf
-mfufq hpmmb sfkk b kjf bmc gvqs zpv
-mfufq hpmmb hjuf zpv vo
-mfufq hpmmb kfs zpv cpxm
-mfufq hpmmb qvm bqpvmc bmc cftfqs zpv
-mfufq hpmmb nblf zpv dqz
-mfufq hpmmb tbz hppcazf
-mfufq hpmmb sfkk b kjf bmc gvqs zpv
+xf\qf(mp(tsqamhfqt(sp(kpuf)zpv(lmpx(sgf(qvkft(amc(tp(cp(j(/cp(j')a(evkk(dpnnjsnfms\t(xgas(j\n(sgjmljmh(pe)zpv(xpvkcm\s(hfs(sgjt(eqpn(amz(psgfq(hvz)j(ivts(xamma(sfkk(zpv(gpx(j\n(effkjmh)hpssa(nalf(zpv(vmcfqtsamc)mfufq(hpmma(hjuf(zpv(vo)mfufq(hpmma(kfs(zpv(cpxm)mfufq(hpmma(qvm(aqpvmc(amc(cftfqs(zpv)mfufq(hpmma(nalf(zpv(dqz)mfufq(hpmma(taz(hppcazf)mfufq(hpmma(sfkk(a(kjf(amc(gvqs(zpv)xf\uf(lmpxm(fadg(psgfq(epq(tp(kpmh)zpvq(gfaqs\t(affm(adgjmh,(avs(zpv\qf(spp(tgz(sp(taz(js(/taz(js')jmtjcf,(xf(apsg(lmpx(xgas\t(affm(hpjmh(pm(/hpjmh(pm')xf(lmpx(sgf(hanf(amc(xf\qf(hpmma(okaz(js)amc(je(zpv(atl(nf(gpx(j\n(effkjmh)cpm\s(sfkk(nf(zpv\qf(spp(akjmc(sp(tff)mfufq(hpmma(hjuf(zpv(vo)mfufq(hpmma(kfs(zpv(cpxm)mfufq(hpmma(qvm(aqpvmc(amc(cftfqs(zpv)mfufq(hpmma(nalf(zpv(dqz)mfufq(hpmma(taz(hppcazf)mfufq(hpmma(sfkk(a(kjf(amc(gvqs(zpv)mfufq(hpmma(hjuf(zpv(vo)mfufq(hpmma(kfs(zpv(cpxm)mfufq(hpmma(qvm(aqpvmc(amc(cftfqs(zpv)mfufq(hpmma(nalf(zpv(dqz)mfufq(hpmma(taz(hppcazf)mfufq(hpmma(sfkk(a(kjf(amc(gvqs(zpv)xf\uf(lmpxm(fadg(psgfq(epq(tp(kpmh)zpvq(gfaqs\t(affm(adgjmh,(avs(zpv\qf(spp(tgz(sp(taz(js(/sp(taz(js')jmtjcf,(xf(apsg(lmpx(xgas\t(affm(hpjmh(pm(/hpjmh(pm')xf(lmpx(sgf(hanf(amc(xf\qf(hpmma(okaz(js)j(ivts(xamma(sfkk(zpv(gpx(j\n(effkjmh)hpssa(nalf(zpv(vmcfqtsamc)mfufq(hpmma(hjuf(zpv(vo)mfufq(hpmma(kfs(zpv(cpxm)mfufq(hpmma(qvm(aqpvmc(amc(cftfqs(zpv)mfufq(hpmma(nalf(zpv(dqz)mfufq(hpmma(taz(hppcazf)mfufq(hpmma(sfkk(a(kjf(amc(gvqs(zpv)mfufq(hpmma(hjuf(zpv(vo)mfufq(hpmma(kfs(zpv(cpxm)mfufq(hpmma(qvm(aqpvmc(amc(cftfqs(zpv)mfufq(hpmma(nalf(zpv(dqz)mfufq(hpmma(taz(hppcazf)mfufq(hpmma(sfkk(a(kjf(amc(gvqs(zpv)mfufq(hpmma(hjuf(zpv(vo)mfufq(hpmma(kfs(zpv(cpxm)mfufq(hpmma(qvm(aqpvmc(amc(cftfqs(zpv)mfufq(hpmma(nalf(zpv(dqz)mfufq(hpmma(taz(hppcazf)mfufq(hpmma(sfkk(a(kjf(amc(gvqs(zpv
