+++ lab_1/texts/task1/key.txt
@@ -1,1 +1,1 @@
-badcfehgjilknmporqtsvuxwzy
+badcfehgjilknmporqtsvuxwzyBADCFEHGJILKNMPORQTSVUXWZY1032547698.,!?\"\\/'() \n
